Extracting code for swipe to edit
